+++ psutil/__init__.py
@@ -2049,8 +2049,6 @@ def disk_io_counters(perdisk=False):
     executed first otherwise this function won't find any disk.
     """
     rawdict = _psplatform.disk_io_counters()
-    if not rawdict:
-        raise RuntimeError("couldn't find any physical disk")
     nt = getattr(_psplatform, "sdiskio", _common.sdiskio)
     if perdisk:
         for disk, fields in rawdict.items():