+++ scripts/who.py
@@ -9,10 +9,10 @@ A clone of 'who' command; print information about users who are
 currently logged in.
 
 $ python scripts/who.py
-giampaolo       tty7            2014-02-23 17:25  (:0)
-giampaolo       pts/7           2014-02-24 18:25  (:192.168.1.56)
-giampaolo       pts/8           2014-02-24 18:25  (:0)
-giampaolo       pts/9           2014-02-27 01:32  (:0)
+giampaolo   tty7     2014-02-23 17:25  (:0)             upstart
+giampaolo   pts/7    2014-02-24 18:25  (:192.168.1.56)  sshd
+giampaolo   pts/8    2014-02-24 18:25  (:0)             upstart
+giampaolo   pts/9    2014-02-27 01:32  (:0)             upstart
 """
 
 from datetime import datetime
@@ -23,11 +23,14 @@ import psutil
 def main():
     users = psutil.users()
     for user in users:
-        print("%-15s %-15s %s  (%s)" % (
+        proc_name = psutil.Process(user.pid).name() if user.pid else ""
+        print("%-12s %-10s %s  (%s) %10s" % (
             user.name,
             user.terminal or '-',
             datetime.fromtimestamp(user.started).strftime("%Y-%m-%d %H:%M"),
-            user.host))
+            user.host,
+            proc_name
+        ))
 
 
 if __name__ == '__main__':