+++ HISTORY.rst
@@ -15,6 +15,8 @@
 
 - 1014_: Linux can mask legitimate ENOENT exceptions as NoSuchProcess.
 - 1016_: disk_io_counters() raises RuntimeError on a system with no disks.
+- 1017_: net_io_counters() raises RuntimeError on a system with no network
+  cards installed.
 
 *2017-04-10*